requirements.txt
#works with Python 3.7.10
catboost==0.26
fastai==2.1.10
lightgbm==3.2.1
numpy==1.19.2
pandas==1.1.5
scikit-learn==0.24.1
scipy==1.6.2
sklearn-contrib-py-earth==0.1.0
torch==1.8.1
xgboost==1.4.0